MyApp/aws-compliance-scanner-B/fastapi-app/requirements.txt
# AWS Compliance Scanner — FastAPI Dependencies
# Approach B: Step Functions Orchestration

fastapi==0.109.2
uvicorn[standard]==0.27.1
pydantic==2.6.1
pydantic-settings==2.1.0

# AWS SDK
boto3==1.34.50
botocore==1.34.50

# Observability
opentelemetry-api==1.22.0
opentelemetry-sdk==1.22.0
opentelemetry-exporter-otlp-proto-grpc==1.22.0
opentelemetry-instrumentation-fastapi==0.43b0
opentelemetry-instrumentation-botocore==0.43b0
prometheus-fastapi-instrumentator==6.1.0

# Testing
pytest==8.0.0
pytest-asyncio==0.23.4
httpx==0.26.0
moto==5.0.1
